% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha2crit}
\alias{alpha2crit}
\title{Alpha to Critical Values}
\usage{
alpha2crit(
  alpha = c(0.001, 0.01, 0.05),
  dist = "z",
  two.tailed = TRUE,
  right.tail = TRUE,
  ...
)
}
\arguments{
\item{alpha}{Numeric vector.
Significance level
\eqn{\left( \alpha \right)} .
By default,
\code{alpha} is set to conventional
significance levels
\code{alpha = c(0.001, 0.01, 0.05)}.}

\item{dist}{Character string.
\code{dist = "z"} for the standard normal distribution.
\code{dist = "t"} for the t distribution.
\code{dist = "F"} for the F distribution.
\code{dist = "chisq"} for the chi-square distribution.}

\item{two.tailed}{Logical.
If \code{TRUE}, two-tailed alpha.
If \code{FALSE}, one-tailed alpha.
Ignored if \code{dist = "F"} or \code{dist = "chisq"}
as both tests are one-tailed using the right tail.}

\item{right.tail}{Logical.
If \code{TRUE}, right tail (positive critical value).
If \code{FALSE}, left tail (negative critical value).
Ignored if \code{two.tailed = TRUE}.
Ignored if \code{dist = "F"} or \code{dist = "chisq"}
as both tests are one-tailed using the right tail.}

\item{...}{Degrees of freedom.
\code{df} for \code{dist = "t"} and \code{dist = "chisq"}.
\code{df1} and \code{df2} for \code{dist = "F"}.}
}
\value{
Returns
\eqn{z}, \eqn{t}, \eqn{\chi^2}, or \eqn{F}
critical value/s
associated with
the specified significance level/s
\eqn{\left( \alpha \right)}.
The results are sorted from smallest to largest.
}
\description{
Calculates the \eqn{z}, \eqn{t}, \eqn{\chi^2}, or \eqn{F}
critical value/s
of confidence limits
associated with the specified significance level/s
\eqn{\left( \alpha \right)} .
}
\examples{
# z two-tailed
## vector
alpha2crit(alpha = c(0.001, 0.01, 0.05))
## single numeric value
alpha2crit(alpha = 0.05)
# t two-tailed
## vector
alpha2crit(alpha = c(0.001, 0.01, 0.05), dist = "t", df = 1000)
## single numeric value
alpha2crit(alpha = 0.05, , dist = "t", df = 1000)
# vector
## two-tailed
### z
alpha2crit(
  alpha = 0.05,
  dist = "z"
)
### t
alpha2crit(
  alpha = 0.05,
  dist = "t",
  df = 5
)
## one-tailed
### right
#### z
alpha2crit(
  alpha = 0.05,
  dist = "z",
  two.tailed = FALSE
)
#### t
alpha2crit(
  alpha = 0.05,
  dist = "t",
  two.tailed = FALSE,
  df = 5
)
#### F
alpha2crit(
  alpha = 0.05,
  dist = "F",
  two.tailed = FALSE,
  df1 = 5,
  df2 = 5
)
#### chi-square
alpha2crit(
  alpha = 0.05,
  dist = "chisq",
  two.tailed = FALSE,
  df = 5
)
### left
#### z
alpha2crit(
  alpha = 0.05,
  dist = "z",
  two.tailed = FALSE,
  right.tail = FALSE
)
#### t
alpha2crit(
  alpha = 0.05,
  dist = "t",
  two.tailed = FALSE,
  right.tail = FALSE,
  df = 5
)
}
\references{
\href{https://en.wikipedia.org/wiki/Statistical_significance}{Wikipedia: Statistical significance}

\href{https://en.wikipedia.org/wiki/Confidence_interval}{Wikipedia: Confidence interval}
}
\seealso{
Other alpha functions: 
\code{\link{alpha2prob}()},
\code{\link{ci2crit}()},
\code{\link{ci2prob}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{alpha functions}
\keyword{alpha}
