% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha2plot}
\alias{alpha2plot}
\title{Alpha to Plot}
\usage{
alpha2plot(
  alpha = 0.05,
  ci = FALSE,
  dist = "z",
  two.tailed = TRUE,
  right.tail = TRUE,
  statistic = NULL,
  ...
)
}
\arguments{
\item{alpha}{Numeric.
Significance level
\eqn{\left( \alpha \right)} .}

\item{ci}{Logical.
\code{ci = FALSE} by default.
If \code{TRUE}, plot confidence interval.
If \code{FALSE}, plot alpha level.}

\item{dist}{Character string.
\code{dist = "z"} for the standard normal distribution.
\code{dist = "t"} for the t distribution.
\code{dist = "F"} for the F distribution.
\code{dist = "chisq"} for the chi-square distribution.}

\item{two.tailed}{Logical.
If \code{TRUE}, two-tailed alpha.
If \code{FALSE}, one-tailed alpha.
Ignored if \code{dist = "F"} or \code{dist = "chisq"}
as both tests are one-tailed using the right tail.}

\item{right.tail}{Logical.
If \code{TRUE}, right tail (positive critical value).
If \code{FALSE}, left tail (negative critical value).
Ignored if \code{two.tailed = TRUE}.
Ignored if \code{dist = "F"} or \code{dist = "chisq"}
as both tests are one-tailed using the right tail.}

\item{statistic}{Numeric.
Test statistic.
This is an optional argument.}

\item{...}{Degrees of freedom.
\code{df} for \code{dist = "t"} and \code{dist = "chisq"}.
\code{df1} and \code{df2} for \code{dist = "F"}.}
}
\description{
Alpha to Plot
}
\examples{
# ci = FALSE
# two-tailed
alpha2plot(
  alpha = 0.05,
  dist = "z",
  statistic = 3
)
alpha2plot(
  alpha = 0.05,
  dist = "t",
  df = 5
)
# one-tailed
alpha2plot(
  alpha = 0.05,
  two.tailed = FALSE,
  right.tail = TRUE
)
alpha2plot(
  alpha = 0.05,
  two.tailed = FALSE,
  right.tail = FALSE
)
alpha2plot(
  alpha = 0.05,
  dist = "t",
  two.tailed = FALSE,
  right.tail = TRUE,
  df = 5
)
alpha2plot(
  alpha = 0.05,
  dist = "t",
  two.tailed = FALSE,
  right.tail = FALSE,
  df = 5
)
alpha2plot(
  alpha = 0.05,
  dist = "F",
  df1 = 10,
  df2 = 3
)
alpha2plot(
  alpha = 0.05,
  dist = "chisq",
  df = 3
)
# ci = TRUE
# two-tailed
alpha2plot(
  alpha = 0.05,
  ci = TRUE,
  dist = "z",
  statistic = 3
)
alpha2plot(
  alpha = 0.05,
  ci = TRUE,
  dist = "t",
  df = 5
)
# one-tailed
alpha2plot(
  alpha = 0.05,
  ci = TRUE,
  two.tailed = FALSE,
  right.tail = TRUE
)
alpha2plot(
  alpha = 0.05,
  ci = TRUE,
  two.tailed = FALSE,
  right.tail = FALSE
)
alpha2plot(
  alpha = 0.05,
  ci = TRUE,
  dist = "t",
  two.tailed = FALSE,
  right.tail = TRUE,
  df = 5
)
alpha2plot(
  alpha = 0.05,
  ci = TRUE,
  dist = "t",
  two.tailed = FALSE,
  right.tail = FALSE,
  df = 5
)
alpha2plot(
  alpha = 0.05,
  ci = TRUE,
  dist = "F",
  df1 = 10,
  df2 = 3
)
alpha2plot(
  alpha = 0.05,
  ci = TRUE,
  dist = "chisq",
  df = 3
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
