% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha2prob}
\alias{alpha2prob}
\title{Alpha to Probabilities}
\usage{
alpha2prob(alpha = c(0.001, 0.01, 0.05))
}
\arguments{
\item{alpha}{Numeric vector.
Alpha level.
By default,
\code{alpha} is set to conventional
alpha levels
\code{alpha = c(0.001, 0.01, 0.05)}.}
}
\value{
Returns
probabilities associated with specified alpha levels.
The results are sorted from smallest to largest.
}
\description{
Calculates the probabilities of confidence limits
associated with specified alpha levels
for a two-tailed test.
}
\examples{
# vector
alpha2prob(alpha = c(0.001, 0.01, 0.05))
# single numeric value
alpha2prob(alpha = 0.05)
}
\seealso{
Other alpha functions: 
\code{\link{ci2prob}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{alpha functions}
\keyword{alpha}
