% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{bca}
\alias{bca}
\title{Confidence Interval - Bias Corrected and Accelerated}
\usage{
bca(
  thetahat_star,
  thetahat,
  data,
  fitFUN,
  alpha = c(0.001, 0.01, 0.05),
  wald = FALSE,
  theta_null = 0,
  distribution = "z",
  df,
  eval = FALSE,
  theta = 0,
  ...
)
}
\arguments{
\item{thetahat_star}{Numeric vector.
The bootstrap sampling distribution (\eqn{\hat{\theta}^{*}}).}

\item{thetahat}{Numeric.
Parameter estimate
(\eqn{\hat{\theta}}).}

\item{data}{Vector, matrix, or data frame.
Sample data.}

\item{fitFUN}{Function.
Fit function to use on \code{data}.
The first argument should correspond to \code{data}.
Other arguments can be passed to \code{fitFUN}
using \code{...}.
\code{fitFUN} should return a single value.}

\item{alpha}{Numeric vector.
Alpha level.
By default,
\code{alpha} is set to conventional
alpha levels
\code{alpha = c(0.001, 0.01, 0.05)}.}

\item{wald}{Logical.
If \code{TRUE},
calculates the square root of the Wald test statistic and p-value.
If \code{FALSE},
returns \code{statistic = NA}
and
\code{p = NA}.
If \code{FALSE},
the arguments
\code{theta_null},
\code{distribution},
and
\code{df}
are ignored.}

\item{theta_null}{Numeric.
Hypothesized value of \code{theta}
(\eqn{\theta_{0}}).
Set to zero by default.}

\item{distribution}{Character string.
\code{distribution = "z"} for the standard normal distribution.
\code{distribution = "t"} for the t distribution.}

\item{df}{Numeric.
Degrees of freedom (df) if \code{distribution = "t"}.
Ignored if \code{distribution = "z"}.}

\item{eval}{Logical.
Evaluate confidence intervals using
\code{\link[=zero_hit]{zero_hit()}},
\code{\link[=theta_hit]{theta_hit()}},
\code{\link[=len]{len()}},
and
\code{\link[=shape]{shape()}}.}

\item{theta}{Numeric.
Population parameter
(\eqn{\theta}).}

\item{...}{Arguments to pass to \code{fitFUN}.}
}
\value{
Returns a vector with the following elements:
\describe{
\item{statistic}{Square root of Wald test statistic. \code{NA} if \code{wald = FALSE}.}
\item{p}{p-value. \code{NA} if \code{wald = FALSE}.}
\item{se}{Standard error of thetahat_star (\eqn{\mathrm{se} \left( \hat{\theta}^{*} \right)}), i.e., standard deviation of thetahat_star (\eqn{\mathrm{sd} \left( \hat{\theta}^{*} \right)}).}
\item{ci_}{Estimated bias corrected and accelerated confidence limits corresponding to alpha from the bootstrap sampling distribution thetahat_star (\eqn{\hat{\theta}^{*}}).}
}
If \code{eval = TRUE},
also returns
\describe{
\item{zero_hit_}{Logical. Tests if confidence interval contains zero.}
\item{theta_hit_}{Logical. Tests if confidence interval contains theta.}
\item{length_}{Length of confidence interval.}
\item{shape_}{Shape of confidence interval.}
}
}
\description{
Calculates bias corrected and accelerated confidence intervals.
}
\examples{
#############################################################
# Generate sample data from a normal distribution
# with mu = 0 and sigma^2 = 1.
#############################################################
set.seed(42)
n <- 10000
x <- rnorm(n = n)

#############################################################
# Estimate the population mean mu using the sample mean xbar.
#############################################################
# Parameter estimate xbar
thetahat <- mean(x)
thetahat
# Closed form solution for the standard error of the mean
se_thetahat <- sd(x) / sqrt(n)
se_thetahat

#############################################################
# Generate B = 2000 nonparametric bootstrap samples.
#############################################################
x_star <- nb(
  data = x,
  B = 2000L,
  par = FALSE,
  ncores = NULL
)

#############################################################
# Estimate xbar for each of the nonparametric bootstrap samples.
# The B = 2000 xbars form
# the empirical sampling distribution thetahat_star.
#############################################################
thetahat_star <- lapply(
  X = x_star,
  FUN = mean
)
thetahat_star <- as.vector(
  do.call(
    what = "rbind",
    args = thetahat_star
  )
)
hist(thetahat_star)
summary(thetahat_star)

#############################################################
# Generate percentile confidence intervals
# for alpha = c(0.001, 0.01, 0.05).
#############################################################
# With square root of Wald test statistic and p-value
# using sd(thetahat_star) as se_thetahat
bca(
  thetahat_star = thetahat_star,
  wald = TRUE,
  thetahat = thetahat,
  data = x,
  fitFUN = mean
)

# Without square root of Wald test statistic and p-value
bca(
  thetahat_star = thetahat_star,
  thetahat = thetahat,
  data = x,
  fitFUN = mean
)
}
\references{
Efron, B., & Tibshirani, R. J. (1993).
An introduction to the bootstrap. New York, N.Y: Chapman & Hall.
}
\seealso{
Other bootstrap confidence interval functions: 
\code{\link{bc}()},
\code{\link{pc}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap confidence interval functions}
\keyword{confidence}
\keyword{interval}
