% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{ci2crit}
\alias{ci2crit}
\title{Confidence Intervals to Critical Values}
\usage{
ci2crit(
  ci = c(0.999, 0.99, 0.95),
  dist = "z",
  two.tailed = TRUE,
  right.tail = TRUE,
  ...
)
}
\arguments{
\item{ci}{Numeric vector.
Confidence interval.
By default,
\code{ci} is set to conventional
confidence intervals
\code{ci = c(0.999, 0.99, 0.95)}.}

\item{dist}{Character string.
\code{dist = "z"} for the standard normal distribution.
\code{dist = "t"} for the t distribution.
\code{dist = "F"} for the F distribution.
\code{dist = "chisq"} for the chi-square distribution.}

\item{two.tailed}{Logical.
If \code{TRUE}, two-tailed alpha.
If \code{FALSE}, one-tailed alpha.
Ignored if \code{dist = "F"} or \code{dist = "chisq"}
as both tests are one-tailed using the right tail.}

\item{right.tail}{Logical.
If \code{TRUE}, right tail (positive critical value).
If \code{FALSE}, left tail (negative critical value).
Ignored if \code{two.tailed = TRUE}.
Ignored if \code{dist = "F"} or \code{dist = "chisq"}
as both tests are one-tailed using the right tail.}

\item{...}{Degrees of freedom.
\code{df} for \code{dist = "t"} and \code{dist = "chisq"}.
\code{df1} and \code{df2} for \code{dist = "F"}.}
}
\value{
Returns
\eqn{z}, \eqn{t}, \eqn{\chi^2}, or \eqn{F}
critical value/s
associated with
the specified confidence interval/s.
The results are sorted from smallest to largest.
}
\description{
Calculates the \eqn{z}, \eqn{t}, \eqn{\chi^2}, or \eqn{F}
critical value/s
of confidence limits
associated with the specified confidence interval/s.
}
\examples{
# vector
## two-tailed
### z
ci2crit(
  ci = 0.95,
  dist = "z"
)
### t
ci2crit(
  ci = 0.95,
  dist = "t",
  df = 5
)
## one-tailed
### right
#### z
ci2crit(
  ci = 0.95,
  dist = "z",
  two.tailed = FALSE
)
#### t
ci2crit(
  ci = 0.95,
  dist = "t",
  two.tailed = FALSE,
  df = 5
)
#### F
ci2crit(
  ci = 0.95,
  dist = "F",
  two.tailed = FALSE,
  df1 = 5,
  df2 = 5
)
#### chi-square
ci2crit(
  ci = 0.95,
  dist = "chisq",
  two.tailed = FALSE,
  df = 5
)
### left
#### z
ci2crit(
  ci = 0.95,
  dist = "z",
  two.tailed = FALSE,
  right.tail = FALSE
)
#### t
ci2crit(
  ci = 0.95,
  dist = "t",
  two.tailed = FALSE,
  right.tail = FALSE,
  df = 5
)
}
\references{
\href{https://en.wikipedia.org/wiki/Statistical_significance}{Wikipedia: Statistical significance}

\href{https://en.wikipedia.org/wiki/Confidence_interval}{Wikipedia: Confidence interval}
}
\seealso{
Other alpha functions: 
\code{\link{alpha2crit}()},
\code{\link{alpha2prob}()},
\code{\link{ci2prob}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{alpha functions}
\keyword{alpha}
