% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci2prob}
\alias{ci2prob}
\title{Confidence Intervals to Probabilities}
\usage{
ci2prob(ci = c(0.999, 0.99, 0.95))
}
\arguments{
\item{ci}{Numeric vector.
Confidence interval.
By default,
\code{ci} is set to conventional
confidence intervals
\code{ci = c(0.999, 0.99, 0.95)}.}
}
\value{
Returns
probabilities associated with specified confidence intervals.
The results are sorted from smallest to biggest.
}
\description{
Calculates the probabilities of confidence limits
associated with specified confidence intervals
for a two-tailed test.
}
\examples{
# vector
ci2prob(ci = c(0.999, 0.99, 0.95))
# single numeric value
ci2prob(ci = 0.95)
}
\author{
Ivan Jacob Agaloos Pesigan
}
