% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{.pb_mvn}
\alias{.pb_mvn}
\title{Parametric Bootstrap
(Multivariate Normal)
from
\eqn{\boldsymbol{\hat{\Sigma}}}
and
\eqn{\boldsymbol{\hat{\mu}}}}
\usage{
.pb_mvn(n, Sigmahat, muhat, B = 2000L, par = FALSE, ncores = NULL)
}
\arguments{
\item{n}{Integer.
Sample size.}

\item{Sigmahat}{Matrix.
Estimated variance-covariance matrix
from the original sample data.}

\item{muhat}{Vector.
Estimated mean vector from the original sample data.}

\item{B}{Integer. Number of bootstrap samples.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}
}
\value{
Returns a list of parametric bootstrap samples.
}
\description{
Generates \code{B} number of parametric bootstrap
samples from the original sample data.
Data is generated from a multivariate normal distribution
using the estimated variance-covariance matrix and mean vector.
}
\examples{
B <- 5L
Sigmahat <- matrix(
  data = c(
    82.37344,
    70.55922,
    17.83930,
    70.55922,
    112.57145,
    -75.98558,
    17.83930,
    -75.98558,
    338.46263
  ),
  nrow = 3
)
muhat <- c(
  108.3060,
  105.3324,
  103.4009
)
.pb_mvn(
  n = 5,
  Sigmahat = Sigmahat,
  muhat = muhat,
  B = B
)
}
\references{
\href{https://en.wikipedia.org/wiki/Bootstrapping_(statistics)}{Wikipedia: Bootstrapping (statistics)}
}
\seealso{
Other bootstrap functions: 
\code{\link{nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap functions}
\keyword{bootstraping}
