% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jack.R
\name{jack}
\alias{jack}
\title{Jackknife}
\usage{
jack(data, par = FALSE, ncores = NULL)
}
\arguments{
\item{data}{Vector, matrix, or data frame.
Sample data.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}
}
\value{
Returns a list of jackknife samples
of length \code{n},
where \code{n} is the sample size of \code{data}.
Each item in the list,
will have a sample size of
\code{n - 1}.
}
\description{
Generates jackknife samples.
}
\details{
\deqn{
  \mathbf{x}_i
  =
  \left\{
    x_1,
    x_2,
    x_3,
    \dots,
    x_n
  \right\}
}
for \eqn{i = \left\{ 1, 2, 3, \dots, n \right\}}
the \eqn{i}th jackknife sample
consists of the original sample data
with the \eqn{i}th observation removed.
}
\examples{
n <- 5
# vector
data_vector <- rnorm(n = n)
jack(
  data = data_vector
)
# matrix
X <- rnorm(n = n)
Y <- rnorm(n = n)
Z <- rnorm(n = n)
data_matrix <- cbind(X, Y, Z)
jack(
  data = data_matrix
)
# data frame
data_dataframe <- data.frame(X, Y, Z)
jack(
  data = data_dataframe
)
}
\references{
\href{https://en.wikipedia.org/wiki/Jackknife_resampling}{Wikipedia: Jackknife resampling}
}
\seealso{
Other jackknife functions: 
\code{\link{jack_hat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{jackknife functions}
\keyword{jackknife}
