% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jack.R
\name{jack_hat}
\alias{jack_hat}
\title{Jackknife Estimates}
\usage{
jack_hat(thetahat_star, thetahat)
}
\arguments{
\item{thetahat_star}{Numeric vector.
Jackknife sampling distribution,
that is,
the sampling distribution of \code{thetahat}
estimated for each \code{i} jackknife sample.
\eqn{
    \hat{\theta}_{\left( 1 \right)},
    \hat{\theta}_{\left( 2 \right)},
    \hat{\theta}_{\left( 3 \right)},
    \dots,
    \hat{\theta}_{\left( n \right)}
  } .}

\item{thetahat}{Numeric.
Parameter estimate
\eqn{\left( \hat{\theta} \right)}
from the original sample data.}
}
\value{
Returns a vector with the following elements:
\describe{
\item{mean}{Mean of \code{thetahat_star} \eqn{\left( \hat{\theta}_{\left( \cdot \right) } \right)}.}
\item{bias}{Jackknife estimate of bias \eqn{\left( \hat{\mathrm{bias}}_{\mathrm{jack}} \left( \theta \right) \right)}.}
\item{se}{Jackknife estimate of standard error \eqn{\left( \hat{\mathrm{se}}_{\mathrm{jack}} \left( \hat{\theta} \right) \right)}.}
\item{thetahat_jack}{Bias-corrected jackknife estimate \eqn{\left( \hat{\theta}_{\mathrm{jack}} \right)}.}
}
}
\description{
Calculates jackknife estimates.
}
\details{
The jackknife estimate of bias is given by
\deqn{
  \hat{\mathrm{bias}}_{\mathrm{jack}}
  \left(
    \theta
  \right)
  =
  \left(
    n - 1
  \right)
  \left(
    \hat{\theta}_{
      \left(
        \cdot
      \right)
    }
    -
    \hat{\theta}
  \right)
}

where

\deqn{
  \hat{\theta}_{
    \left(
      \cdot
    \right)
  }
  =
  \frac{1}{n}
  \sum_{i = 1}^{n}
  \hat{\theta}_{
    \left(
      i
    \right)
  } .
}

The jackknife estimate of standard error
is given by

\deqn{
  \hat{\mathrm{se}}_{\mathrm{jack}}
  \left(
    \hat{\theta}
  \right)
  =
  \sqrt{
    \frac{n - 1}{n}
    \sum_{i = 1}^{n}
    \left(
      \hat{\theta}_{\left( i \right)}
      -
      \hat{\theta}_{\left( \cdot \right)}
    \right)^2
  } .
}

The bias-corrected jackknife estimate
is given by

\deqn{
  \hat{\theta}_{\mathrm{jack}}
  =
  \hat{\theta}
  -
  \hat{\mathrm{bias}}_{\mathrm{jack}}
  \left(
    \theta
  \right)
  =
  n
  \hat{\theta}
  -
  \left(
    n - 1
  \right)
  \hat{\theta}_{
    \left(
      \cdot
    \right)
  } .
}
}
\examples{
n <- 100
# vector
x <- rnorm(n = n)
thetahat <- mean(x)
jack_samples <- jack(
  data = x
)
thetahat_star <- sapply(
  X = jack_samples,
  FUN = mean
)
hist(thetahat_star)
jack_hat(
  thetahat_star = thetahat_star,
  thetahat = thetahat
)
}
\references{
\href{https://en.wikipedia.org/wiki/Jackknife_resampling}{Wikipedia: Jackknife resampling}
}
\seealso{
Other jackknife functions: 
\code{\link{jack}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{jackknife functions}
\keyword{jackknife}
