% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{len}
\alias{len}
\title{Confidence Interval - Length}
\usage{
len(thetahat_lo, thetahat_up)
}
\arguments{
\item{thetahat_lo}{Numeric.
Lower limit of the estimated confidence interval
(\eqn{\hat{\theta}_{\mathrm{lo}}}).}

\item{thetahat_up}{Numeric.
Upper limit of the estimated confidence interval
(\eqn{\hat{\theta}_{\mathrm{up}}}).}
}
\description{
Calculates confidence interval length.
}
\details{
The confidence interval length is given by
\deqn{
  \mathrm{confidence \ interval \ length}
  =
  \hat{\theta}_{\mathrm{up}}
  -
  \hat{\theta}_{\mathrm{lo}}
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
