% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lsat.R
\docType{data}
\name{lsat}
\alias{lsat}
\title{Law School Data}
\format{
A matrix with 15 cases and 2 variables:
\describe{
\item{LSAT}{Law School Admission Test.}
\item{GPA}{Grade Point Average.}
}
}
\source{
Bradley Efron and Robert J. Tibshirani, 1993, "An Introduction to the Bootstrap",
Table 3.1
}
\usage{
lsat
}
\description{
Law School Data
}
\keyword{datasets}
