% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc}
\alias{mc}
\title{Monte Carlo}
\usage{
mc(thetahat, covhat_thetahat, R = 20000L)
}
\arguments{
\item{thetahat}{Numeric vector.
Parameter estimates
\eqn{\boldsymbol{\hat{\theta}}}.}

\item{covhat_thetahat}{Symmetric matrix.
Estimated sampling variances and covariances of
parameter estimates
\eqn{\hat{\mathrm{Cov}} \left( \boldsymbol{\hat{\theta}} \right)}.}

\item{R}{Integer.
Number of Monte Carlo replications.}
}
\description{
Generates \code{R} number of parameter estimates
from a vector of \code{thetahat}
\eqn{\left( \boldsymbol{\hat{\theta} }\right)}
and
sampling variances and covariances of \code{thetahat}
\eqn{\hat{\mathrm{Cov}} \left( \boldsymbol{\hat{\theta}} \right)}.
If length of \code{thetahat} is greater than 1,
a multivariate normal distribution is assumed.
If length of \code{thetahat} is equal 1,
a univariate normal distribution is assumed.
}
\examples{
R <- 20000L
# length 1
thetahat <- 100
covhat_thetahat <- 1.5
mc_star_length_1 <- mc(
  thetahat = thetahat,
  covhat_thetahat = covhat_thetahat,
  R = R
)
head(mc_star_length_1)
# length greater than 1
alphahat <- 0.3386
betahat <- 0.4510
alphahat_betahat <- alphahat * betahat
varhat_alphahat <- 0.1224^2
varhat_betahat <- 0.1460^2
thetahat <- c(
  alphahat,
  betahat
)
covhat_thetahat <- matrix(
  data = c(
    varhat_alphahat,
    0.00,
    0.00,
    varhat_betahat
  ),
  ncol = 2
)
mc_star_length_2 <- mc(
  thetahat = thetahat,
  covhat_thetahat = covhat_thetahat,
  R = R
)
head(mc_star_length_2)
alphahat_betahat_star <- mc_star_length_2[, 1] * mc_star_length_2[, 2]
hist(
  alphahat_betahat_star,
  main = expression(
    paste(
      "Histogram of ",
      hat(alpha),
      hat(beta),
      "*"
    )
  ),
  xlab = expression(
    paste(
      hat(alpha),
      hat(beta),
      "*"
    )
  )
)
wald(
  thetahat = alphahat_betahat,
  sehat_thetahat = sd(alphahat_betahat_star)
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
