% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{nb}
\alias{nb}
\title{Nonparametric Bootstrap}
\usage{
nb(data, B = 2000L, par = FALSE, ncores = NULL)
}
\arguments{
\item{data}{Vector, matrix or data frame.
Sample data to bootstrap.}

\item{B}{Integer. Number of bootstrap samples.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}
}
\value{
Returns a list of nonparametric bootstrap samples.
}
\description{
Generates \code{B} number of nonparametric bootstrap
samples from the original sample data.
}
\examples{
B <- 5L
# vector
data_vector <- rnorm(n = 5)
nb(
  data = data_vector,
  B = B
)
# matrix
X <- rnorm(n = 5)
M <- rnorm(n = 5)
Y <- rnorm(n = 5)
data_matrix <- cbind(X, M, Y)
nb(
  data = data_matrix,
  B = B
)
# data frame
data_dataframe <- data.frame(X, M, Y)
nb(
  data = data_dataframe,
  B = B
)
}
\references{
\href{https://en.wikipedia.org/wiki/Bootstrapping_(statistics)}{Wikipedia: Bootstrapping (statistics)}
}
\seealso{
Other bootstrap functions: 
\code{\link{.pb_mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap functions}
\keyword{bootstraping}
