% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{pb_univ}
\alias{pb_univ}
\title{Parametric Bootstrap (Univariate)}
\usage{
pb_univ(rFUN = rnorm, n, B = 2000L, par = FALSE, ncores = NULL, ...)
}
\arguments{
\item{rFUN}{Function.
Data generating function to generate univariate data.}

\item{n}{Integer.
Sample size.}

\item{B}{Integer. Number of bootstrap samples.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}

\item{...}{Arguments to pass to rFUN.}
}
\description{
Parametric Bootstrap (Univariate)
}
\examples{
pb_univ(
  rFUN = rexp,
  n = 100,
  rate = 1
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
