% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{pbmvn}
\alias{pbmvn}
\title{Parametric Bootstrap
(Multivariate Normal)
from
\eqn{
  \boldsymbol{
    \hat{
      \mu
    }
  }
  \left(
    \boldsymbol{
      \hat{
        \theta
      }
    }
  \right)
  %(\#eq:boot-pb-mvn-mu)
}
and
\eqn{
  \boldsymbol{
    \hat{
      \Sigma
    }
  }
  \left(
    \boldsymbol{
      \hat{
        \theta
      }
    }
  \right)
  %(\#eq:boot-pb-mvn-Sigma)
} .}
\usage{
pbmvn(
  n,
  muhatthetahat,
  Sigmahatthetahat,
  tol = 1e-06,
  empirical = FALSE,
  B = 2000L,
  par = FALSE,
  ncores = NULL,
  mc = TRUE,
  lb = FALSE,
  cl_eval = FALSE,
  cl_export = FALSE,
  cl_expr,
  cl_vars
)
}
\arguments{
\item{n}{Integer.
Sample size.}

\item{muhatthetahat}{Vector.
Mean vector as a function of estimated parameters
or the fitted model-implied mean vector
\eqn{
    \boldsymbol{
      \hat{
        \mu
      }
    }
    \left(
      \boldsymbol{
        \hat{
          \theta
        }
      }
    \right)
    %(\#eq:boot-pb-mvn-mu)
  } .}

\item{Sigmahatthetahat}{Matrix.
Variance-covariance matrix as a function of estimated parameters
or the fitted model-implied variance-covariance matrix
\eqn{
    \boldsymbol{
      \hat{
        \Sigma
      }
    }
    \left(
      \boldsymbol{
        \hat{
          \theta
        }
      }
    \right)
    %(\#eq:boot-pb-mvn-Sigma)
  } .}

\item{tol}{Numeric.
Tolerance (relative to largest variance)
for numerical lack of positive-definiteness in \code{Sigma}.}

\item{empirical}{Logical.
If \code{TRUE}, \code{mu} and \code{Sigma} specify the empirical
not population mean and covariance matrix.}

\item{B}{Integer.
Number of bootstrap samples.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.
If \code{FALSE}, use \code{\link[=lapply]{lapply()}}.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{mc}{Logical.
If \code{TRUE}, use \code{\link[parallel:mclapply]{parallel::mclapply()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}} or \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
Ignored if \code{par = FALSE}.}

\item{lb}{Logical.
If \code{TRUE} use \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}}.
Ignored if \code{par = FALSE} and \code{mc = TRUE}.}

\item{cl_eval}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}} using \code{cl_expr}.
Ignored if \code{mc = TRUE}.}

\item{cl_export}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterExport()}} using \code{cl_vars}.
Ignored if \code{mc = TRUE}.}

\item{cl_expr}{Expression.
Expression passed to \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}}
Ignored if \code{mc = TRUE}.}

\item{cl_vars}{Character vector.
Names of objects to pass to \code{\link[parallel:clusterApply]{parallel::clusterExport()}}
Ignored if \code{mc = TRUE}.}
}
\value{
Returns a list of length \code{B} of parametric bootstrap samples.
}
\description{
Generates \code{B} number of parametric bootstrap
samples using estimated parameters
from the original sample \code{data}.
\code{data} is referred to as the empirical distribution
with the following distributional assumption
\deqn{
  \hat{
    F
  }_{
    \mathcal{
      N
    }_{k}
    \left(
      \boldsymbol{
        \hat{
          \mu
        }
      }
      \left(
        \boldsymbol{
          \hat{
            \theta
          }
        }
      \right) ,
      \boldsymbol{
        \hat{
          \Sigma
        }
      }
      \left(
        \boldsymbol{
          \hat{
            \theta
          }
        }
      \right)
    \right)
  } .
  %(\#eq:boot-pb-mvn)
}
Bootstrap samples are generated from a multivariate normal distribution
using the fitted model-implied mean vector
and variance-covariance matrix.
}
\details{
For more details and examples see the following vignettes:

\href{https://jeksterslabds.github.io/jeksterslabRboot/articles/notes/notes_intro_nb.html}{Notes: Intro to NB}

\href{https://jeksterslabds.github.io/jeksterslabRboot/articles/notes/notes_intro_pb.html}{Notes: Intro to PB}
}
\examples{
B <- 5L
Sigmahatthetahat <- matrix(
  data = c(
    82.37344,
    70.55922,
    17.83930,
    70.55922,
    112.57145,
    -75.98558,
    17.83930,
    -75.98558,
    338.46263
  ),
  nrow = 3
)
muhatthetahat <- c(
  108.3060,
  105.3324,
  103.4009
)
Xstar <- pbmvn(
  n = 5,
  Sigmahatthetahat = Sigmahatthetahat,
  muhatthetahat = muhatthetahat,
  B = B
)
str(Xstar)
}
\references{
Efron, B., & Tibshirani, R. J. (1993).
\emph{An introduction to the bootstrap}.
New York, N.Y: Chapman & Hall.

\href{https://en.wikipedia.org/wiki/Bootstrapping_(statistics)}{Wikipedia: Bootstrapping (statistics)}
}
\seealso{
Other bootstrap functions: 
\code{\link{nb}()},
\code{\link{pbuniv}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap functions}
\keyword{bootstrap}
