% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{pbuniv}
\alias{pbuniv}
\title{Parametric Bootstrap (Univariate)}
\usage{
pbuniv(
  n,
  rFUN = rnorm,
  ...,
  B = 2000L,
  par = FALSE,
  ncores = NULL,
  mc = TRUE,
  lb = FALSE,
  cl_eval = FALSE,
  cl_export = FALSE,
  cl_expr,
  cl_vars,
  rbind = NULL
)
}
\arguments{
\item{n}{Integer.
Sample size.}

\item{rFUN}{Function.
Data generating function to generate univariate data.}

\item{...}{Arguments to pass to \code{rFUN}.}

\item{B}{Integer.
Number of bootstrap samples.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.
If \code{FALSE}, use \code{\link[=lapply]{lapply()}}.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{mc}{Logical.
If \code{TRUE}, use \code{\link[parallel:mclapply]{parallel::mclapply()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}} or \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
Ignored if \code{par = FALSE}.}

\item{lb}{Logical.
If \code{TRUE} use \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}}.
Ignored if \code{par = FALSE} and \code{mc = TRUE}.}

\item{cl_eval}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}} using \code{cl_expr}.
Ignored if \code{mc = TRUE}.}

\item{cl_export}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterExport()}} using \code{cl_vars}.
Ignored if \code{mc = TRUE}.}

\item{cl_expr}{Expression.
Expression passed to \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}}
Ignored if \code{mc = TRUE}.}

\item{cl_vars}{Character vector.
Names of objects to pass to \code{\link[parallel:clusterApply]{parallel::clusterExport()}}
Ignored if \code{mc = TRUE}.}

\item{rbind}{NULL or logical.
If \code{rbind = NULL},
returns the list produced.
If \code{TRUE}, uses \code{\link[=rbind]{rbind()}}
to bind the rows of the list produced.
If \code{FALSE}, uses \code{\link[=cbind]{cbind()}}
to bind the columns of the list produced.
Test that each element of the output list
has the appropriate dimensions for binding
before using this option.}
}
\value{
Returns a list of length \code{B} of parametric bootstrap samples.
}
\description{
Generates \code{B} number of parametric bootstrap
samples from estimated parameters of original univariate sample \code{data}.
\code{data} is referred to as the empirical distribution
\deqn{
  \hat{
    F
  }_{
    \mathcal{
      Distribution
    }
  }
}.
The default distribution is
\deqn{
  \mathcal{
    N
  }
  \left(
    \hat{
      \mu
    },
    \hat{
      \sigma
    }^2
  \right)
  %(\#eq:boot-pb-norm)
}.
The univariate distribution and parameters used in the
data generating process can be specified using \code{rFUN} and \code{...}.
}
\details{
For more details and examples see the following vignettes:

\href{https://jeksterslabds.github.io/jeksterslabRboot/articles/notes/notes_intro_nb.html}{Notes: Intro to NB}

\href{https://jeksterslabds.github.io/jeksterslabRboot/articles/notes/notes_intro_pb.html}{Notes: Intro to PB}
}
\examples{
n <- 5
B <- 5
# normal distribution---------------------------------------------------------
mu <- 100
sigma2 <- 225
sigma <- sqrt(sigma2)
x <- rnorm(
  n = n,
  mean = mu,
  sd = sigma
)
muhatthetahat <- mean(x)
Sigmahatthetahat <- sd(x)
xstar <- pbuniv(
  n = n,
  rFUN = rnorm,
  B = B,
  mean = muhatthetahat,
  sd = Sigmahatthetahat
)
str(xstar)
# binomial distribution-------------------------------------------------------
n_trials <- 1
p <- 0.50
x <- rbinom(
  n = n,
  size = n_trials,
  prob = p
)
phat <- mean(x) / n_trials
xstar <- pbuniv(
  n = n,
  rFUN = rbinom,
  B = B,
  size = n_trials,
  prob = phat
)
str(xstar)
}
\references{
Efron, B., & Tibshirani, R. J. (1993).
\emph{An introduction to the bootstrap}.
New York, N.Y: Chapman & Hall.

\href{https://en.wikipedia.org/wiki/Bootstrapping_(statistics)}{Wikipedia: Bootstrapping (statistics)}
}
\seealso{
Other bootstrap functions: 
\code{\link{nb}()},
\code{\link{pbmvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap functions}
\keyword{bootstrap}
