% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit.R
\name{theta_hit}
\alias{theta_hit}
\title{Confidence Interval - Theta Hit}
\usage{
theta_hit(thetahat_lo, theta, thetahat_up)
}
\arguments{
\item{thetahat_lo}{Numeric.
Lower limit of the estimated confidence interval
(\eqn{\hat{\theta}_{\mathrm{lo}}}).}

\item{theta}{Numeric.
Population parameter
(\eqn{\theta}).}

\item{thetahat_up}{Numeric.
Upper limit of the estimated confidence interval
(\eqn{\hat{\theta}_{\mathrm{up}}}).}
}
\value{
Returns
\code{TRUE} if \code{theta} (\eqn{\theta}) is between the interval
\code{thetahat_lo}
(\eqn{\hat{\theta}_{\mathrm{lo}}})
to
\code{thetahat_up}
(\eqn{\hat{\theta}_{\mathrm{up}}}).
Returns
\code{FALSE} if \code{theta} (\eqn{\theta}) is outside the interval
\code{thetahat_lo}
(\eqn{\hat{\theta}_{\mathrm{lo}}})
to
\code{thetahat_up}
(\eqn{\hat{\theta}_{\mathrm{up}}}).
}
\description{
Confidence Interval - Theta Hit
}
\examples{
# FALSE
theta_hit(thetahat_lo = 1, theta = 0, thetahat_up = 2)
# TRUE
theta_hit(thetahat_lo = -1, theta = 0, thetahat_up = 1)
}
\seealso{
Other confidence interval hit functions: 
\code{\link{zero_hit}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{confidence interval hit functions}
\keyword{confidence}
\keyword{interval}
