% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{wald_test}
\alias{wald_test}
\title{Wald Test Statistic for a Single Parameter}
\usage{
wald_test(thetahat, varhat, null = 0)
}
\arguments{
\item{thetahat}{Numeric.
Parameter estimate
\eqn{\left( \hat{\theta} \right)}.}

\item{varhat}{Numeric.
Estimated variance of \code{thetahat}
\eqn{\left( \widehat{\mathrm{Var}} \left( \hat{\theta} \right) \right)}.}

\item{null}{Numeric.
Hypothesized value of \code{theta}
\eqn{\left( \theta_{0} \right)}.
Set to zero by default.}
}
\value{
Returns a vector with the following elements:
\describe{
\item{statistic}{Wald test statistic.}
\item{p}{p-value.}
}
}
\description{
Calculates the Wald test statistic for a single parameter.
}
\details{
The Wald test statistic for a single parameter is calculated
as follows:
\deqn{
    W
    =
    \frac{
      \left(
        \hat{\theta} - \theta_0
      \right)^2
    }
    {
      \widehat{\mathrm{Var}}
      \left(
        \hat{\theta}
      \right)
    }
  }
The associated \code{p}-value
from the \code{chi-square} distribution is calculated
using \code{\link[=pchisq]{pchisq()}} with \code{df = 1}.
}
\examples{
thetahat <- 0.860
sehat <- 0.409
varhat <- sehat^2
wald_test(
  thetahat = thetahat,
  varhat = varhat
)
}
\references{
\href{https://en.wikipedia.org/wiki/Wald_test}{Wikipedia: Wald test}
}
\seealso{
Other Wald test functions: 
\code{\link{sqrt_wald_test}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Wald test functions}
\keyword{Wald}
\keyword{test}
