% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha2prob}
\alias{alpha2prob}
\title{Alpha to Probabilities}
\usage{
alpha2prob(alpha = c(0.001, 0.01, 0.05))
}
\arguments{
\item{alpha}{Numeric vector.
Significance level
\eqn{\left( \alpha \right)} .
By default,
\code{alpha} is set to conventional
significance levels
\code{alpha = c(0.001, 0.01, 0.05)}.}
}
\value{
Returns
probabilities associated with
the specified significance level/s
\eqn{\left( \alpha \right)}
for a two-tailed test.
The results are sorted from smallest to largest.
}
\description{
Calculates the cumulative probabilities of confidence limits
associated with the specified significance level/s
\eqn{\left( \alpha \right)}
for a two-tailed test.
}
\examples{
# vector
alpha2prob(alpha = c(0.001, 0.01, 0.05))
# single numeric value
alpha2prob(alpha = 0.05)
}
\references{
\href{https://en.wikipedia.org/wiki/Statistical_significance}{Wikipedia: Statistical significance}

\href{https://en.wikipedia.org/wiki/Confidence_interval}{Wikipedia: Confidence interval}
}
\seealso{
Other alpha functions: 
\code{\link{alpha2crit}()},
\code{\link{ci2crit}()},
\code{\link{ci2prob}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{alpha functions}
\keyword{alpha}
