% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{ci_eval}
\alias{ci_eval}
\title{Confidence Interval Evaluation}
\usage{
ci_eval(ci, thetahat, theta = 0, label = NULL)
}
\arguments{
\item{ci}{Vector.
Confidence intervals sorted from smallest to largest.
The length should be even.
The first and the last element correspond to the widest confidence interval.
The second and the second to the last element correspond to the second widest confidence interval.
And so on and so forth.}

\item{thetahat}{Numeric.
Parameter estimate
\eqn{\left( \hat{\theta} \right)}.}

\item{theta}{Numeric.
Population parameter
\eqn{\left( \theta \right)}.}

\item{label}{Vector.
Vector used to label results.
If not provided defaults to \code{label = 1:(length(ci)/2)}.}
}
\value{
Returns a vector with the following elements:
\describe{
\item{zero_hit_}{Logical. Tests if confidence interval contains zero.}
\item{theta_hit_}{Logical. Tests if confidence interval contains theta.}
\item{length_}{Length of confidence interval.}
\item{shape_}{Shape of confidence interval.}
}
}
\description{
Evaluates confidence interval using
\code{\link[=zero_hit]{zero_hit()}},
\code{\link[=theta_hit]{theta_hit()}},
\code{\link[=len]{len()}},
and
\code{\link[=shape]{shape()}}.
}
\examples{
ci <- c(
  98.04786,
  98.38773,
  98.68060,
  100.5447,
  100.8375,
  101.1774
)
thetahat <- 99.6126336
theta <- 100
label <- c(
  0.001,
  0.01,
  0.05
)
ci_eval(
  ci = ci,
  thetahat = thetahat,
  theta = theta,
  label = label
)
}
\seealso{
Other confidence interval evaluation functions: 
\code{\link{len}()},
\code{\link{shape}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{confidence interval evaluation functions}
\keyword{confidence}
\keyword{interval}
