% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{.pb_mvn}
\alias{.pb_mvn}
\title{Parametric Bootstrap
(Multivariate Normal)
from
\eqn{\boldsymbol{\hat{\Sigma}}}
and
\eqn{\boldsymbol{\hat{\mu}}}}
\usage{
.pb_mvn(n, Sigmahat, muhat, B = 2000L, par = FALSE, ncores = NULL)
}
\arguments{
\item{n}{Integer.
Sample size.}

\item{Sigmahat}{Matrix.
Estimated variance-covariance matrix
from the original sample data.}

\item{muhat}{Vector.
Estimated mean vector from the original sample data.}

\item{B}{Integer. Number of bootstrap samples.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}
}
\value{
Returns a list of length \code{B} of parametric bootstrap samples.
}
\description{
Generates \code{B} number of parametric bootstrap
samples from the original sample \code{data}.
\code{data} is referred to as the empirical distribution
\eqn{
  \hat{F}_{
    \mathrm{MVN}
    \left(
      \boldsymbol{\hat{\Sigma}},
      \boldsymbol{\hat{\mu}}
    \right)
  }
} .
Data is generated from a multivariate normal distribution
using the estimated variance-covariance matrix
\eqn{\boldsymbol{\hat{\Sigma}}}
and
mean vector
\eqn{\boldsymbol{\hat{\mu}}}
.
}
\details{
For more details and examples see the following vignettes:

\href{https://jeksterslabds.github.io/jeksterslabRboot/articles/notes/notes_intro_nb.html}{Notes: Introduction to Nonparametric Bootstrapping}

\href{https://jeksterslabds.github.io/jeksterslabRboot/articles/notes/notes_intro_pb.html}{Notes: Introduction to Parametric Bootstrapping}
}
\examples{
B <- 5L
Sigmahat <- matrix(
  data = c(
    82.37344,
    70.55922,
    17.83930,
    70.55922,
    112.57145,
    -75.98558,
    17.83930,
    -75.98558,
    338.46263
  ),
  nrow = 3
)
muhat <- c(
  108.3060,
  105.3324,
  103.4009
)
.pb_mvn(
  n = 5,
  Sigmahat = Sigmahat,
  muhat = muhat,
  B = B
)
}
\references{
Efron, B., & Tibshirani, R. J. (1993).
An introduction to the bootstrap. New York, N.Y: Chapman & Hall.

\href{https://en.wikipedia.org/wiki/Bootstrapping_(statistics)}{Wikipedia: Bootstrapping (statistics)}
}
\seealso{
Other bootstrap functions: 
\code{\link{nb}()},
\code{\link{pb_mvn}()},
\code{\link{pb_univ}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap functions}
\keyword{bootstrap}
