% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jack.R
\name{jack}
\alias{jack}
\title{Jackknife}
\usage{
jack(
  data,
  par = FALSE,
  ncores = NULL,
  mc = TRUE,
  lb = FALSE,
  cl_eval = FALSE,
  cl_export = FALSE,
  cl_expr,
  cl_vars
)
}
\arguments{
\item{data}{Vector, matrix, or data frame.
Sample data.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.
If \code{FALSE}, use \code{\link[=lapply]{lapply()}}.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{mc}{Logical.
If \code{TRUE}, use \code{\link[parallel:mclapply]{parallel::mclapply()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}} or \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
Ignored if \code{par = FALSE}.}

\item{lb}{Logical.
If \code{TRUE} use \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}}.
Ignored if \code{par = FALSE} and \code{mc = TRUE}.}

\item{cl_eval}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}} using \code{cl_expr}.
Ignored if \code{mc = TRUE}.}

\item{cl_export}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterExport()}} using \code{cl_vars}.
Ignored if \code{mc = TRUE}.}

\item{cl_expr}{Expression.
Expression passed to \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}}
Ignored if \code{mc = TRUE}.}

\item{cl_vars}{Character vector.
Names of objects to pass to \code{\link[parallel:clusterApply]{parallel::clusterExport()}}
Ignored if \code{mc = TRUE}.}
}
\value{
Returns a list of jackknife samples
of length \code{n},
where \code{n} is the sample size of \code{data}.
Each item in the list,
will have a sample size of
\code{n - 1}.
}
\description{
Generates jackknife samples.
}
\details{
\deqn{
  \mathbf{
    x
  }_i
  =
  \left\{
    x_1,
    x_2,
    x_3,
    \dots,
    x_n
  \right\}
}
for
\eqn{
  i
  =
  \left\{
    1,
    2,
    3,
    \dots,
    n
  \right\}
}
the \eqn{i}th jackknife sample
consists of the original sample data
with the \eqn{i}th observation removed.
}
\examples{
n <- 5
#################################
# vector
#################################
x <- rnorm(n = n)
xstar <- jack(
  data = x
)
str(xstar)
#################################
# matrix
#################################
x1 <- rnorm(n = n)
x2 <- rnorm(n = n)
x3 <- rnorm(n = n)
X <- cbind(x1, x2, x3)
Xstar <- jack(
  data = X
)
str(Xstar)
#################################
# data frame
#################################
X <- as.data.frame(X)
Xstar <- jack(
  data = X
)
str(Xstar)
}
\references{
\href{https://en.wikipedia.org/wiki/Jackknife_resampling}{Wikipedia: Jackknife resampling}
}
\seealso{
Other jackknife functions: 
\code{\link{jack_hat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{jackknife functions}
\keyword{jackknife}
