% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jack.R
\name{jack_hat}
\alias{jack_hat}
\title{Jackknife Estimates}
\usage{
jack_hat(
  thetahatstarjack,
  thetahat,
  alpha = c(0.001, 0.01, 0.05),
  eval = FALSE,
  theta = 0
)
}
\arguments{
\item{thetahatstarjack}{Numeric vector.
Jackknife sampling distribution,
that is,
the sampling distribution of \code{thetahat}
estimated for each \code{i} jackknife sample.
\eqn{
    \hat{\theta}_{\left( 1 \right)},
    \hat{\theta}_{\left( 2 \right)},
    \hat{\theta}_{\left( 3 \right)},
    \dots,
    \hat{\theta}_{\left( n \right)}
  } .}

\item{thetahat}{Numeric.
Parameter estimate
\eqn{\left( \hat{\theta} \right)}
from the original sample data.}

\item{alpha}{Numeric vector.
Significance level
\eqn{\left( \alpha \right)} .
By default,
\code{alpha} is set to conventional
significance levels
\code{alpha = c(0.001, 0.01, 0.05)}.}

\item{eval}{Logical.
Evaluate confidence intervals using
\code{\link[=zero_hit]{zero_hit()}},
\code{\link[=theta_hit]{theta_hit()}},
\code{\link[=len]{len()}},
and
\code{\link[=shape]{shape()}}.}

\item{theta}{Numeric.
Population parameter
\eqn{\left( \theta \right)}.}
}
\value{
Returns a list with the following elements:
\describe{
\item{hat}{Jackknife estimates.}
\item{ps}{Pseudo-values.}
\item{ci}{Confidence intervals using pseudo-values.}
}
The first list element \code{hat} contains the following:
\describe{
\item{mean}{Mean of \code{thetahatstarjack} \eqn{\left( \hat{\theta}_{\left( \cdot \right) } \right)}.}
\item{bias}{Jackknife estimate of bias \eqn{\left( \widehat{\mathrm{bias}}_{\mathrm{jack}} \left( \theta \right) \right)}.}
\item{se}{Jackknife estimate of standard error \eqn{\left( \widehat{\mathrm{se}}_{\mathrm{jack}} \left( \hat{\theta} \right) \right)}.}
\item{thetahatjack}{Bias-corrected jackknife estimate \eqn{\left( \hat{\theta}_{\mathrm{jack}} \right)}.}
}
}
\description{
Calculates jackknife estimates.
}
\details{
The jackknife estimate of bias is given by
\deqn{
  \widehat{\mathrm{bias}}_{\mathrm{jack}}
  \left(
    \theta
  \right)
  =
  \left(
    n - 1
  \right)
  \left(
    \hat{\theta}_{
      \left(
        \cdot
      \right)
    }
    -
    \hat{\theta}
  \right)
}

where

\deqn{
  \hat{\theta}_{
    \left(
      \cdot
    \right)
  }
  =
  \frac{1}{n}
  \sum_{i = 1}^{n}
  \hat{\theta}_{
    \left(
      i
    \right)
  } .
}

The jackknife estimate of standard error
is given by

\deqn{
  \widehat{\mathrm{se}}_{\mathrm{jack}}
  \left(
    \hat{\theta}
  \right)
  =
  \sqrt{
    \frac{n - 1}{n}
    \sum_{i = 1}^{n}
    \left(
      \hat{\theta}_{\left( i \right)}
      -
      \hat{\theta}_{\left( \cdot \right)}
    \right)^2
  } .
}

The bias-corrected jackknife estimate
is given by

\deqn{
  \hat{\theta}_{\mathrm{jack}}
  =
  \hat{\theta}
  -
  \hat{\mathrm{bias}}_{\mathrm{jack}}
  \left(
    \theta
  \right)
  =
  n
  \hat{\theta}
  -
  \left(
    n - 1
  \right)
  \hat{\theta}_{
    \left(
      \cdot
    \right)
  } .
}

Pseudo-values can be computed using

\deqn{
  \tilde{\theta}_{i}
  =
  n \hat{\theta}
  -
  \left(
    n - 1
  \right)
  \hat{\theta}_{\left( i \right)} .
}

The standard error can be estimated using the pseudo-values

\deqn{
  \widehat{\mathrm{se}}_{\mathrm{jack}}
  \left(
    \tilde{\theta}
  \right)
  =
  \sqrt{
    \sum_{i = 1}^{n}
    \frac{
      \left(
        \tilde{\theta}_{i}
        -
        \tilde{\theta}
      \right)^2
    }
    {
      \left(
        n
        -
        1
      \right)
      n
    }
  }
}

where

\deqn{
  \tilde{\theta}
  =
  \frac{1}{n}
  \sum_{i = 1}^{n}
  \tilde{\theta}_{i} .
}

An interval can be generated using

\deqn{
  \tilde{\theta}
  \pm
  t_{\frac{\alpha}{2}}
  \times
  \widehat{\mathrm{se}}_{\mathrm{jack}}
    \left(
      \tilde{\theta}
    \right)
}
with degrees of freedom \eqn{\nu = n - 1}.
}
\examples{
n <- 100
x <- rnorm(n = n)
thetahat <- mean(x)
xstar <- jack(
  data = x
)
thetahatstarjack <- sapply(
  X = xstar,
  FUN = mean
)
str(xstar, list.len = 6)
hist(
  thetahatstarjack,
  main = expression(
    paste(
      "Histogram of ",
      hat(theta),
      "*"
    )
  ),
  xlab = expression(
    paste(
      hat(theta),
      "*"
    )
  )
)
jack_hat(
  thetahatstarjack = thetahatstarjack,
  thetahat = thetahat
)
}
\references{
\href{https://en.wikipedia.org/wiki/Jackknife_resampling}{Wikipedia: Jackknife resampling}
}
\seealso{
Other jackknife functions: 
\code{\link{jack}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{jackknife functions}
\keyword{jackknife}
