% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{len}
\alias{len}
\title{Confidence Interval - Length}
\usage{
len(lo, up)
}
\arguments{
\item{lo}{Numeric.
Lower limit of the estimated confidence interval
\eqn{\left( \hat{\theta}_{\mathrm{lo}} \right)}.}

\item{up}{Numeric.
Upper limit of the estimated confidence interval
\eqn{\left( \hat{\theta}_{\mathrm{up}} \right)}.}
}
\description{
Calculates confidence interval length.
}
\details{
The confidence interval length is given by
\deqn{
  \mathrm{confidence \ interval \ length}
  =
  \hat{\theta}_{\mathrm{up}}
  -
  \hat{\theta}_{\mathrm{lo}}
}
}
\seealso{
Other confidence interval evaluation functions: 
\code{\link{shape}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{confidence interval evaluation functions}
\keyword{confidence}
\keyword{interval}
