% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{nb}
\alias{nb}
\title{Nonparametric Bootstrap}
\usage{
nb(data, B = 2000L, par = FALSE, ncores = NULL)
}
\arguments{
\item{data}{Vector, matrix or data frame.
Sample data to bootstrap.
The empirical distribution \eqn{\hat{F}}.}

\item{B}{Integer. Number of bootstrap samples.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}
}
\value{
Returns a list of length \code{B} of nonparametric bootstrap samples.
}
\description{
Generates \code{B} number of nonparametric bootstrap
samples from the original sample \code{data}.
\code{data} is referred to as the empirical distribution \eqn{\hat{F}}.
}
\details{
For more details and examples see the following vignettes:

\href{https://jeksterslabds.github.io/jeksterslabRboot/articles/notes/notes_intro_nb.html}{Notes: Introduction to Nonparametric Bootstrapping}

\href{https://jeksterslabds.github.io/jeksterslabRboot/articles/notes/notes_intro_pb.html}{Notes: Introduction to Parametric Bootstrapping}
}
\examples{
B <- 5L
n <- 5
#################################
# vector
#################################
x <- rnorm(n = n)
xstar <- nb(
  data = x,
  B = B
)
str(xstar)
#################################
# matrix
#################################
x1 <- rnorm(n = n)
x2 <- rnorm(n = n)
x3 <- rnorm(n = n)
X <- cbind(x1, x2, x3)
Xstar <- nb(
  data = X,
  B = B
)
str(Xstar)
#################################
# data frame
#################################
X <- as.data.frame(X)
Xstar <- nb(
  data = X,
  B = B
)
str(Xstar)
}
\references{
Efron, B., & Tibshirani, R. J. (1993).
\emph{An introduction to the bootstrap}.
New York, N.Y: Chapman & Hall.

\href{https://en.wikipedia.org/wiki/Bootstrapping_(statistics)}{Wikipedia: Bootstrapping (statistics)}
}
\seealso{
Other bootstrap functions: 
\code{\link{.pb_mvn}()},
\code{\link{pb_mvn}()},
\code{\link{pb_univ}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap functions}
\keyword{bootstrap}
