% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{nb}
\alias{nb}
\title{Nonparametric Bootstrap}
\usage{
nb(data, B = 2000L, par = FALSE, ncores = NULL)
}
\arguments{
\item{data}{Vector, matrix or data frame.
Sample data to bootstrap.
The empirical distribution \eqn{\hat{F}}.}

\item{B}{Integer. Number of bootstrap samples.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}
}
\value{
Returns a list of length \code{B} of nonparametric bootstrap samples.
}
\description{
Generates \code{B} number of nonparametric bootstrap
samples from the original sample \code{data}.
\code{data} is referred to as the empirical distribution \eqn{\hat{F}}.
}
\examples{
B <- 5L
n <- 5
# vector
data_vector <- rnorm(n = n)
nb(
  data = data_vector,
  B = B
)
# matrix
X <- rnorm(n = n)
Y <- rnorm(n = n)
Z <- rnorm(n = n)
data_matrix <- cbind(X, Y, Z)
nb(
  data = data_matrix,
  B = B
)
# data frame
data_dataframe <- data.frame(X, Y, Z)
nb(
  data = data_dataframe,
  B = B
)
}
\references{
Efron, B., & Tibshirani, R. J. (1993).
An introduction to the bootstrap. New York, N.Y: Chapman & Hall.

\href{https://en.wikipedia.org/wiki/Bootstrapping_(statistics)}{Wikipedia: Bootstrapping (statistics)}
}
\seealso{
Other bootstrap functions: 
\code{\link{.pb_mvn}()},
\code{\link{pb_mvn}()},
\code{\link{pb_univ}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap functions}
\keyword{bootstrap}
