% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{pb_mvn}
\alias{pb_mvn}
\title{Parametric Bootstrap
(Multivariate Normal)
from Data}
\usage{
pb_mvn(data, B = 2000L, par = FALSE, ncores = NULL)
}
\arguments{
\item{data}{Vector, matrix or data frame.
Sample data to bootstrap.
The empirical distribution \eqn{\hat{F}}.}

\item{B}{Integer. Number of bootstrap samples.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}
}
\value{
Returns a list of length \code{B} of parametric bootstrap samples.
}
\description{
Generates \code{B} number of parametric bootstrap
samples from the original sample \code{data}.
\code{data} is referred to as the empirical distribution
\eqn{
  \hat{F}_{
    \mathrm{MVN}
    \left(
      \boldsymbol{\hat{\Sigma}},
      \boldsymbol{\hat{\mu}}
    \right)
  }
} .
Data is generated from a multivariate normal distribution
using the estimated variance-covariance matrix
\eqn{\boldsymbol{\hat{\Sigma}}}
and
mean vector
\eqn{\boldsymbol{\hat{\mu}}}
.
}
\examples{
data <- matrix(
  data = c(
    121.03482,
    99.74874,
    101.44951,
    114.31910,
    104.97803,
    118.31483,
    101.39711,
    90.20035,
    105.31186,
    111.43784,
    95.68517,
    109.52585,
    105.11345,
    128.37871,
    78.30152
  ),
  nrow = 5
)
pb_mvn(
  data = data,
  B = 5L
)
}
\references{
Efron, B., & Tibshirani, R. J. (1993).
An introduction to the bootstrap. New York, N.Y: Chapman & Hall.

\href{https://en.wikipedia.org/wiki/Bootstrapping_(statistics)}{Wikipedia: Bootstrapping (statistics)}
}
\seealso{
Other bootstrap functions: 
\code{\link{.pb_mvn}()},
\code{\link{nb}()},
\code{\link{pb_univ}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap functions}
\keyword{bootstrap}
