% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{pb_univ}
\alias{pb_univ}
\title{Parametric Bootstrap (Univariate)}
\usage{
pb_univ(rFUN = rnorm, n, B = 2000L, par = FALSE, ncores = NULL, ...)
}
\arguments{
\item{rFUN}{Function.
Data generating function to generate univariate data.}

\item{n}{Integer.
Sample size.}

\item{B}{Integer. Number of bootstrap samples.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}

\item{...}{Arguments to pass to rFUN.}
}
\value{
Returns a list of length \code{B} of parametric bootstrap samples.
}
\description{
Generates \code{B} number of parametric bootstrap
samples from estimated parameters of original univariate sample \code{data}.
\code{data} is referred to as the empirical distribution
\eqn{\hat{F}_{\mathcal{Distribution}}}.
The default distribution is
\eqn{
  \mathcal{N}
  \left(
    \hat{\mu},
    \hat{\sigma}^2
  \right)
}.
The univariate distribution and parameters used in the
data generating process can be specified using \code{rFUN} and \code{...}.
}
\details{
For more details and examples see the following vignettes:

\href{https://jeksterslabds.github.io/jeksterslabRboot/articles/notes/notes_intro_nb.html}{Notes: Introduction to Nonparametric Bootstrapping}

\href{https://jeksterslabds.github.io/jeksterslabRboot/articles/notes/notes_intro_pb.html}{Notes: Introduction to Parametric Bootstrapping}
}
\examples{
n <- 5
B <- 5
#################################
# normal distribution
#################################
mu <- 100
sigma2 <- 225
sigma <- sqrt(sigma2)
x <- rnorm(
  n = n,
  mean = mu,
  sd = sigma
)
muhat <- mean(x)
sigmahat <- sd(x)
xstar <- pb_univ(
  rFUN = rnorm,
  n = n,
  B = B,
  mean = muhat,
  sd = sigmahat
)
str(xstar)
#################################
# binomial distribution
#################################
n_trials <- 1
p <- 0.50
x <- rbinom(
  n = n,
  size = n_trials,
  prob = p
)
phat <- mean(x) / n_trials
xstar <- pb_univ(
  rFUN = rbinom,
  n = n,
  B = B,
  size = n_trials,
  prob = phat
)
str(xstar)
}
\references{
Efron, B., & Tibshirani, R. J. (1993).
\emph{An introduction to the bootstrap}.
New York, N.Y: Chapman & Hall.

\href{https://en.wikipedia.org/wiki/Bootstrapping_(statistics)}{Wikipedia: Bootstrapping (statistics)}
}
\seealso{
Other bootstrap functions: 
\code{\link{.pb_mvn}()},
\code{\link{nb}()},
\code{\link{pb_mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap functions}
\keyword{bootstrap}
