% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{pc}
\alias{pc}
\title{Confidence Interval - Percentile}
\usage{
pc(
  thetahat_star,
  alpha = c(0.001, 0.01, 0.05),
  wald = FALSE,
  thetahat,
  theta_null = 0,
  distribution = "z",
  df
)
}
\arguments{
\item{thetahat_star}{Numeric vector.
The bootstrap sampling distribution (\eqn{\hat{\theta}^{*}}).}

\item{alpha}{Numeric vector.
Alpha level.
By default,
\code{alpha} is set to conventional
alpha levels
\code{alpha = c(0.001, 0.01, 0.05)}.}

\item{wald}{Logical.
If \code{TRUE},
calculates the square root of the Wald test statistic and p-value.
If \code{FALSE},
returns \code{statistic = NA}
and
\code{p = NA}.
If \code{FALSE},
the arguments
\code{thetahat},
\code{theta_null},
\code{distribution},
and
\code{df}
are ignored.}

\item{thetahat}{Numeric.
Parameter estimate
(\eqn{\hat{\theta}}).}

\item{theta_null}{Numeric.
Hypothesized value of \code{theta}
(\eqn{\theta_{0}}).
Set to zero by default.}

\item{distribution}{Character string.
\code{distribution = "z"} for the standard normal distribution.
\code{distribution = "t"} for the t distribution.}

\item{df}{Numeric.
Degrees of freedom (df) if \code{distribution = "t"}.
Ignored if \code{distribution = "z"}.}
}
\value{
Returns a vector with the following elements:
\describe{
\item{statistic}{Square root of Wald test statistic. \code{NA} if \code{wald = FALSE}.}
\item{p}{p-value. \code{NA} if \code{wald = FALSE}.}
\item{se}{Standard error of thetahat_star (\eqn{\mathrm{se} \left( \hat{\theta}^{*} \right)}), i.e., standard deviation of thetahat_star (\eqn{\mathrm{sd} \left( \hat{\theta}^{*} \right)}).}
\item{ci_*}{Estimated percentile confidence limits corresponding to alpha from the bootstrap sampling distribution thetahat_star (\eqn{\hat{\theta}^{*}}).}
}
}
\description{
Calculates percentile confidence intervals.
}
\details{
The standard error of the bootstrap sampling distribution
\eqn{\hat{\theta}^{*}}
is given by
\deqn{
  \mathrm{se}
  \left(
    \hat{\theta}^{*}
  \right)
  =
  \sqrt{
    \frac{1}{B - 1}
    \sum_{b = 1}^{B}
    \left[
      \hat{\theta}^{*} \left( b \right)
      -
      \hat{\theta}^{*} \left( \cdot \right)
    \right]^2
  }
}
where
the mean of the bootstrap sampling distribution
\eqn{\hat{\theta}^{*}}
is given by
\deqn{
  \hat{\theta}^{*}
  \left(
    \cdot
  \right)
  =
  \frac{1}{n}
  \sum_{b = 1}^{B}
  \hat{\theta}^{*} \left( b \right) .
}

The percentile confidence interval is given by
\deqn{
    \left[
      \hat{\theta}_{\mathrm{lo}},
      \hat{\theta}_{\mathrm{up}}
    \right]
    =
    \left[
      \hat{\theta}^{*}_{\frac{\alpha}{2}},
      \hat{\theta}^{*}_{1 - \frac{\alpha}{2}}
    \right] .
  }
The corresponding percentile value
from the bootstrap sampling distribution \code{thetahat_star}
(\eqn{\hat{\theta}^{*}}) is calculated using
\code{\link[=quantile]{quantile()}}.
}
\examples{
#############################################################
# Generate sample data from a normal distribution
# with mu = 0 and sigma^2 = 1.
#############################################################
set.seed(42)
n <- 10000
x <- rnorm(n = n)

#############################################################
# Estimate the population mean mu using the sample mean xbar.
#############################################################
# Parameter estimate xbar
thetahat <- mean(x)
thetahat
# Closed form solution for the standard error of the mean
se_thetahat <- sd(x) / sqrt(n)
se_thetahat

#############################################################
# Generate B = 2000 nonparametric bootstrap samples.
#############################################################
x_star <- nb(
  data = x,
  B = 2000L,
  par = FALSE,
  ncores = NULL
)

#############################################################
# Estimate xbar for each of the nonparametric bootstrap samples.
# The B = 2000 xbars form
# the empirical sampling distribution thetahat_star.
#############################################################
thetahat_star <- lapply(
  X = x_star,
  FUN = mean
)
thetahat_star <- as.vector(
  do.call(
    what = "rbind",
    args = thetahat_star
  )
)
hist(thetahat_star)
summary(thetahat_star)

#############################################################
# Generate percentile confidence intervals
# for alpha = c(0.001, 0.01, 0.05).
#############################################################
# With square root of Wald test statistic and p-value
# using sd(thetahat_star) as se_thetahat
pc(
  thetahat_star = thetahat_star,
  wald = TRUE,
  thetahat = thetahat
)

# Without square root of Wald test statistic and p-value
pc(
  thetahat_star = thetahat_star
)
}
\references{
Bradley Efron and Robert J. Tibshirani, 1993, "An Introduction to the Bootstrap"
}
\author{
Ivan Jacob Agaloos Pesigan
}
