% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{pc}
\alias{pc}
\title{Confidence Interval - Percentile}
\usage{
pc(
  thetahat_star,
  thetahat,
  alpha = c(0.001, 0.01, 0.05),
  wald = FALSE,
  theta_null = 0,
  distribution = "z",
  df,
  eval = FALSE,
  theta = 0
)
}
\arguments{
\item{thetahat_star}{Numeric vector.
The bootstrap sampling distribution
\eqn{\left( \hat{\theta}^{*} \right)},
that is,
the sampling distribution of \code{thetahat}
estimated for each \code{b} bootstrap sample.
\eqn{
    \hat{\theta}_{\left( 1 \right)},
    \hat{\theta}_{\left( 2 \right)},
    \hat{\theta}_{\left( 3 \right)},
    \hat{\theta}_{\left( b \right)},
    \dots,
    \hat{\theta}_{\left( B \right)}
  } .}

\item{thetahat}{Numeric.
Parameter estimate
\eqn{\left( \hat{\theta} \right)}
from the original sample data.}

\item{alpha}{Numeric vector.
Alpha level.
By default,
\code{alpha} is set to conventional
alpha levels
\code{alpha = c(0.001, 0.01, 0.05)}.}

\item{wald}{Logical.
If \code{TRUE},
calculates the square root of the Wald test statistic and p-value.
The estimated bootstrap standard error is used.
If \code{FALSE},
returns \code{statistic = NA}
and
\code{p = NA}.
If \code{FALSE},
the arguments
\code{theta_null},
\code{distribution},
and
\code{df}
are ignored.}

\item{theta_null}{Numeric.
Hypothesized value of \code{theta}
\eqn{\left( \theta_{0} \right)}.
Set to zero by default.}

\item{distribution}{Character string.
\code{distribution = "z"} for the standard normal distribution.
\code{distribution = "t"} for the t distribution.}

\item{df}{Numeric.
Degrees of freedom (df) if \code{distribution = "t"}.
Ignored if \code{distribution = "z"}.}

\item{eval}{Logical.
Evaluate confidence intervals using
\code{\link[=zero_hit]{zero_hit()}},
\code{\link[=theta_hit]{theta_hit()}},
\code{\link[=len]{len()}},
and
\code{\link[=shape]{shape()}}.}

\item{theta}{Numeric.
Population parameter
\eqn{\left( \theta \right)}.}
}
\value{
Returns a vector with the following elements:
\describe{
\item{statistic}{Square root of Wald test statistic. \code{NA} if \code{wald = FALSE}.}
\item{p}{p-value. \code{NA} if \code{wald = FALSE}.}
\item{se}{Estimated bootstrap standard error \eqn{\left( \hat{\mathrm{se}}_{\mathrm{B}} \left( \hat{\theta} \right) \right)}.}
\item{ci_}{Estimated percentile confidence limits corresponding to alpha from the bootstrap sampling distribution thetahat_star \eqn{\left( \hat{\theta}^{*} \right)}.}
}
If \code{eval = TRUE},
appends the following to the results vector
\describe{
\item{zero_hit_}{Logical. Tests if confidence interval contains zero.}
\item{theta_hit_}{Logical. Tests if confidence interval contains theta.}
\item{length_}{Length of confidence interval.}
\item{shape_}{Shape of confidence interval.}
}
}
\description{
Calculates percentile confidence intervals.
}
\details{
The estimated bootstrap standard error
is given by
\deqn{
  \hat{\mathrm{se}}_{\mathrm{B}}
  \left(
    \hat{\theta}
  \right)
  =
  \sqrt{
    \frac{1}{B - 1}
    \sum_{b = 1}^{B}
    \left[
      \hat{\theta}^{*} \left( b \right)
      -
      \hat{\theta}^{*} \left( \cdot \right)
    \right]^2
  }
}
where
\deqn{
  \hat{\theta}^{*}
  \left(
    \cdot
  \right)
  =
  \frac{1}{B}
  \sum_{b = 1}^{B}
  \hat{\theta}^{*} \left( b \right) .
}

Note that
\eqn{
  \hat{\mathrm{se}}_{\mathrm{B}}
  \left(
    \hat{\theta}
  \right)
}
is the standard deviation of
\eqn{
  \hat{\theta}^{*}
}
and
\eqn{
  \hat{\theta}^{*}
  \left(
    \cdot
  \right)
}
is the mean of
\eqn{
  \hat{\theta}^{*}
} .

The percentile confidence interval is given by
\deqn{
    \left[
      \hat{\theta}_{\mathrm{lo}},
      \hat{\theta}_{\mathrm{up}}
    \right]
    =
    \left[
      \hat{\theta}^{*}_{z_{\left( \frac{\alpha}{2} \right)}},
      \hat{\theta}^{*}_{z_{\left( 1 - \frac{\alpha}{2} \right)}}
    \right] .
  }

For more details and examples see the following vignettes:

\href{https://jeksterslabds.github.io/jeksterslabRboot/articles/notes/notes_intro_nb.html}{Notes: Introduction to Nonparametric Bootstrapping}

\href{https://jeksterslabds.github.io/jeksterslabRboot/articles/notes/notes_intro_pb.html}{Notes: Introduction to Parametric Bootstrapping}
}
\references{
Efron, B., & Tibshirani, R. J. (1993).
An introduction to the bootstrap. New York, N.Y: Chapman & Hall.
}
\seealso{
Other bootstrap confidence interval functions: 
\code{\link{bca}()},
\code{\link{bc}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{bootstrap confidence interval functions}
\keyword{confidence}
\keyword{interval}
