% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{shape}
\alias{shape}
\title{Confidence Interval - Shape}
\usage{
shape(thetahat_lo, thetahat, thetahat_up)
}
\arguments{
\item{thetahat_lo}{Numeric.
Lower limit of the estimated confidence interval
(\eqn{\hat{\theta}_{\mathrm{lo}}}).}

\item{thetahat}{Numeric.
Parameter estimate
(\eqn{\hat{\theta}}).}

\item{thetahat_up}{Numeric.
Upper limit of the estimated confidence interval
(\eqn{\hat{\theta}_{\mathrm{up}}}).}
}
\description{
Calculates confidence interval shape.
}
\details{
The confidence interval shape is given by
\deqn{
  \mathrm{confidence \ interval \ shape}
  =
  \frac{
    \hat{\theta}_{\mathrm{up}} - \hat{\theta}
  }
  {
    \hat{\theta} - \hat{\theta}_{\mathrm{lo}}
  }
}

The shape measures the asymmetry of the confidence interval
around the point estimate \eqn{\hat{\theta}}.
Shape \eqn{> 1.00} is indicative of greater distance between
\eqn{\hat{\theta}_{\mathrm{up}}} and \eqn{\hat{\theta}}
than
\eqn{\hat{\theta}} and \eqn{\hat{\theta}_{\mathrm{lo}}}.
}
\seealso{
Other confidence interval evaluation functions: 
\code{\link{len}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{confidence interval evaluation functions}
\keyword{confidence}
\keyword{interval}
