% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{sqrt_wald_test}
\alias{sqrt_wald_test}
\title{Square Root of the Wald Test Statistic for a Single Parameter}
\usage{
sqrt_wald_test(
  thetahat,
  sehat_thetahat,
  theta_null = 0,
  distribution = "z",
  df
)
}
\arguments{
\item{thetahat}{Numeric.
Parameter estimate
\eqn{\left( \hat{\theta} \right)}.}

\item{sehat_thetahat}{Numeric.
Estimated standard error of \code{thetahat}
\eqn{\left( \hat{\mathrm{se}} \left( \hat{\theta} \right) \right)}.}

\item{theta_null}{Numeric.
Hypothesized value of \code{theta}
\eqn{\left( \theta_{0} \right)}.
Set to zero by default.}

\item{distribution}{Character string.
\code{distribution = "z"} for the standard normal distribution.
\code{distribution = "t"} for the t distribution.}

\item{df}{Numeric.
Degrees of freedom (df) if \code{distribution = "t"}.
Ignored if \code{distribution = "z"}.}
}
\value{
Returns a vector with the following elements:
\describe{
\item{statistic}{Square root of Wald test statistic.}
\item{p}{p-value.}
}
}
\description{
Calculates the square root of the Wald test statistic for a single parameter.
}
\details{
The square root of the Wald test statistic for a single parameter is calculated
as follows:
\deqn{
    \sqrt{W}
    =
    \frac{
      \hat{\theta} - \theta_0
    }
    {
      \hat{\mathrm{se}}
      \left(
        \hat{\theta}
      \right)
    } .
  }
The associated \code{p}-value
from the \code{z} or \code{t} distribution is calculated
using \code{\link[=pnorm]{pnorm()}} or \code{\link[=pt]{pt()}}.
}
\examples{
thetahat <- 0.860
sehat_thetahat <- 0.409
sqrt_wald_test(
  thetahat = thetahat,
  sehat_thetahat = sehat_thetahat
)
sqrt_wald_test(
  thetahat = thetahat,
  sehat_thetahat = sehat_thetahat,
  distribution = "t",
  df = 1000
)
}
\references{
\href{https://en.wikipedia.org/wiki/Wald_test}{Wikipedia: Wald test}
}
\seealso{
Other Wald test functions: 
\code{\link{wald_test}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Wald test functions}
\keyword{Wald}
\keyword{test}
