% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{wald}
\alias{wald}
\title{Confidence Interval - Wald}
\usage{
wald(
  thetahat,
  sehat_thetahat,
  theta_null = 0,
  alpha = c(0.001, 0.01, 0.05),
  distribution = "z",
  df,
  eval = FALSE,
  theta = 0
)
}
\arguments{
\item{thetahat}{Numeric.
Parameter estimate
\eqn{\left( \hat{\theta} \right)}.}

\item{sehat_thetahat}{Numeric.
Estimated standard error of \code{thetahat}
\eqn{\left( \hat{\mathrm{se}} \left( \hat{\theta} \right) \right)}.}

\item{theta_null}{Numeric.
Hypothesized value of \code{theta}
\eqn{\left( \theta_{0} \right)}.
Set to zero by default.}

\item{alpha}{Numeric vector.
Alpha level.
By default,
\code{alpha} is set to conventional
alpha levels
\code{alpha = c(0.001, 0.01, 0.05)}.}

\item{distribution}{Character string.
\code{distribution = "z"} for the standard normal distribution.
\code{distribution = "t"} for the t distribution.}

\item{df}{Numeric.
Degrees of freedom (df) if \code{distribution = "t"}.
Ignored if \code{distribution = "z"}.}

\item{eval}{Logical.
Evaluate confidence intervals using
\code{\link[=zero_hit]{zero_hit()}},
\code{\link[=theta_hit]{theta_hit()}},
\code{\link[=len]{len()}},
and
\code{\link[=shape]{shape()}}.}

\item{theta}{Numeric.
Population parameter
(\eqn{\theta}).}
}
\value{
Returns a vector with the following elements:
\describe{
\item{statistic}{Square root of Wald test statistic.}
\item{p}{p-value.}
\item{se}{Estimated d=standard error of thetahat \eqn{\left( \hat{\mathrm{se}} \left( \hat{\theta} \right) \right)}.}
\item{ci_}{Estimated confidence limits corresponding to alpha.}
}
If \code{eval = TRUE},
also returns
\describe{
\item{zero_hit_}{Logical. Tests if confidence interval contains zero.}
\item{theta_hit_}{Logical. Tests if confidence interval contains theta.}
\item{length_}{Length of confidence interval.}
\item{shape_}{Shape of confidence interval.}
}
}
\description{
Calculates symmetric Wald confidence intervals.
}
\details{
As the sample size approaches infinity \eqn{\left( n \to \infty \right)},
the distribution of \eqn{\hat{\theta}}
approaches the normal distribution
with mean equal to \eqn{\theta}
and variance equal to \eqn{\hat{\mathrm{Var}} \left( \hat{\theta} \right)}
\deqn{
    \hat{\theta}
    \mathrel{\dot\sim}
    \mathcal{N}
    \left(
      \theta,
      \hat{\mathrm{Var}}
      \left(
        \hat{\theta}
      \right)
    \right) .
  }
As such,
\eqn{
    \hat{\theta}
  }
can be expressed in terms of
\eqn{z}-scores
from a standard normal distribution
\deqn{
    \frac{
      \hat{\theta} - \theta
    }
    {
      \hat{\mathrm{se}}
      \left(
        \hat{\theta}
      \right)
    }
    \mathrel{\dot\sim}
    \mathcal{N}
    \left(
      0,
      1
    \right)
  }
where
\eqn{
    \hat{\mathrm{se}}
      \left(
        \hat{\theta}
      \right)
    =
    \sqrt{
      \hat{\mathrm{Var}}
      \left(
        \hat{\theta}
      \right)
    }
  }.

To form a confidence interval around \eqn{\hat{\theta}},
the \eqn{z}-score associated with a particular
alpha level
can be plugged-in the equation below.
\deqn{
    \hat{\theta}
    \pm
    z_{\frac{\alpha}{2}}
    \times
    \hat{\mathrm{se}}
      \left(
        \hat{\theta}
      \right)
  }

Note that this is valid only when \eqn{n \to \infty}.
In finite samples,
this is only an approximation.
Gosset derived a better approximation
in the context of \eqn{\hat{\theta} = \bar{x}}
\deqn{
    \frac{
      \hat{\theta} - \theta
    }
    {
      \hat{\mathrm{se}}
      \left(
        \hat{\theta}
      \right)
    }
    \mathrel{\dot\sim}
    t \left( \nu \right)
  }
where
\eqn{t}
is the Student's
\eqn{t} distribution
and
\eqn{\nu} is the degrees of freedom
\eqn{n - 1}.
As such,
the symmetric Wald confidence interval is given by
\deqn{
    \hat{\theta}
    \pm
    t_{\frac{\alpha}{2}}
    \times
    \hat{\mathrm{se}}
      \left(
        \hat{\theta}
      \right)
  }
from a \eqn{t} distribution with degrees of freedom \eqn{\nu = n - 1}.
Note that in large sample sizes,
\eqn{t} converges to \eqn{z}.
}
\examples{
#############################################################
# Generate sample data from a normal distribution
# with mu = 0 and sigma^2 = 1.
#############################################################
set.seed(42)
n <- 10000
x <- rnorm(n = n)

#############################################################
# Estimate the population mean mu using the sample mean xbar.
#############################################################
# Parameter estimate xbar
thetahat <- mean(x)
thetahat
# Closed form solution for the standard error of the mean
sehat_thetahat <- sd(x) / sqrt(n)
sehat_thetahat

#############################################################
# Generate Wald confidence intervals
# for alpha = c(0.001, 0.01, 0.05).
#############################################################
wald(
  thetahat = thetahat,
  sehat_thetahat = sehat_thetahat
)
}
\references{
\href{https://en.wikipedia.org/wiki/Confidence_interval}{Wikipedia: Confidence interval}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Wald confidence interval functions}
\keyword{Wald}
\keyword{confidence}
\keyword{interval,}
\keyword{test}
