% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{wald_test}
\alias{wald_test}
\title{Wald Test Statistic for a Single Parameter}
\usage{
wald_test(thetahat, var_thetahat, theta_null)
}
\arguments{
\item{thetahat}{Numeric.
Parameter estimate
(\eqn{\hat{\theta}}).}

\item{var_thetahat}{Numeric.
Variance of \code{thetahat}
(\eqn{\mathrm{Var} \left( \hat{\theta} \right)}).}

\item{theta_null}{Numeric.
Hypothesized value of \code{theta}
(\eqn{\theta_{0}}).
Set to zero by default.}
}
\value{
Returns a vector with the following elements:
\describe{
\item{statistic}{Wald test statistic.}
\item{p}{p-value.}
}
}
\description{
Calculates the Wald test statistic for a single parameter.
}
\details{
The Wald test statistic for a single parameter is calculated
as follows:
\deqn{
    W
    =
    \frac{
      \left(
        \hat{\theta} - \theta_0
      \right)^2
    }
    {
      \mathrm{Var}
      \left(
        \hat{\theta}
      \right)
    }
  }
The associated \code{p}-value
from the \code{chi-square} distribution is calculated
using \code{\link[=pchisq]{pchisq()}} with \code{df = 1}.
}
\references{
\href{https://en.wikipedia.org/wiki/Wald_test}{Wikipedia: Wald test}
}
\author{
Ivan Jacob Agaloos Pesigan
}
