% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit.R
\name{zero_hit}
\alias{zero_hit}
\title{Confidence Interval - Zero Hit}
\usage{
zero_hit(thetahat_lo, thetahat_up)
}
\arguments{
\item{thetahat_lo}{Numeric.
Lower limit of the estimated confidence interval
(\eqn{\hat{\theta}_{\mathrm{lo}}}).}

\item{thetahat_up}{Numeric.
Upper limit of the estimated confidence interval
(\eqn{\hat{\theta}_{\mathrm{up}}}).}
}
\value{
Returns
\code{TRUE} if zero is between the interval
\code{thetahat_lo}
(\eqn{\hat{\theta}_{\mathrm{lo}}})
to
\code{thetahat_up}
(\eqn{\hat{\theta}_{\mathrm{up}}}).
Returns
\code{FALSE} if zero is outside the interval
\code{thetahat_lo}
(\eqn{\hat{\theta}_{\mathrm{lo}}})
to
\code{thetahat_up}
(\eqn{\hat{\theta}_{\mathrm{up}}}).
}
\description{
Confidence Interval - Zero Hit
}
\examples{
# FALSE
zero_hit(thetahat_lo = 1, thetahat_up = 2)
# TRUE
zero_hit(thetahat_lo = -1, thetahat_up = 1)
}
\seealso{
Other confidence interval hit functions: 
\code{\link{theta_hit}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{confidence interval hit functions}
\keyword{confidence}
\keyword{interval}
